requirements.txt
Flask
boto3
psycopg2-binary
werkzeug
python-dotenv

